Enable sharing of sockets between different process groups
socket_config instances may specify shared=True, which means that
all process groups having the same socket_config.url (not the same
socket_config) will share the socket.

This is implemented as a metaclass for SocketManager, so that
whole SocketManager instances are shared, not just raw sockets.
The metaclass is basically an implementation of weak references.
The built-in weakref module cannot be used because it doesn't support
objects with sockets inside before Python version 2.4.

The two-level reference counting is neccessary to preserve the
interface (the metaclass is roughly to the SocketManager as the
SocketManager is to the sockets inside).