Verify that key/value pairs are correctly separated.
Test that there is, in fact, an '=' equals sign separating the key and value. If not, there is probably a syntax error in the value somewhere, such as a missing comma. Without this change, the lexer happily uses the next key as a separator, and provided there are enough tokens left you get unexpected results:

     >>> dict_of_key_value_pairs('KEY1=look-ma-no-comma KEY2=ends-with-comma,')
     {'KEY1': 'look-ma-no-comma', '=': ','}

Testing for the equals sign catches the above error.