Let the supervisord daemon log to syslog.
When the options ``stdout_logfile`` or ``stderr_logfile`` are set to
"syslog", the children are logging to the syslog service,
but the daemon is still logging to a file.

By letting the possibility to set the value for ``logfile`` to
"syslog", we can have the same behavior for the daemon.