Redo "Parse env configuration in posix mode"
Fixes #328

This allows (escaped) quotes in the values as well as empty
values.

This was done in pull request #329 but removed as it
broke parsing empty quotes (#873) due to a bug in shlex
(http://bugs.python.org/issue21999). This bug is fixed
so posix mode can be used.

Brings back #329
Partially reverts #880