Revert "Merge pull request #241 from jaraco/log-targets"
The syslog changes introduced in this merge are being reverted on the
Supervisor 3.x branch only.  These changes remain in the 4.x branch.
There are a number of open issues related to syslog support now, and
more work may need to be done here.  We've decided not to release this
in 3.x, but we do intend to include it in Supervisor 4.

This reverts commit 2ab53646eb740a5e594b917e6df3e4a4ce913715, reversing
changes made to 4d4f9c9201f34b2bc5cd553ce612481c43a36d66.

Conflicts:
	docs/configuration.rst
	supervisor/tests/test_dispatchers.py