Add grace time option for child termination
When receiving a termination signal the old behavior was to relay the
signal to the child process and terminating. The problem was that the
child process may keep running and supervisor thinking that it has
ended, because pidproxy ended.

Now the default behavior is to wait until the child process end
before pidproxy terminating. Also an option was added to wait for
grace time seconds the child process termination before sendig a
kill signal.