changed very misleading cmdline option names
In retrospect, picking "--wait" as a command line parameter was not wise;
users would assume it would make pidproxy wait for a subprocess to terminate.
I am changing it with --daemon and --start-wait, in order to make this less misleading.
--start-wait defaults to 60, which is a reasonable amount of time. Hopefully,
typical users would just need to add --daemon and be fairly happy with the result.