Add mode to honor linebreaks in stdout
In order to avoid line breaks in the middle of log lines (which
currently happens when the buffer is flushed) we need a variable length
buffer. This is solved by flushing only up to the next line break and
keeping the remainder of the current buffer until the next flush.

In this commit a config parameter for this is added:
honor_log_linebreaks=true

The default value is false to not break existing behaviour.