print "stayed up for > x seconds" instead of "stayed up for > than x seconds"
Because `for longer than than x seconds` doesn't look nice and is not gramatically correct, well not literally. But the `>` sign is a `greater-than` (or `more-than` or - fitting for our use-case - `longer-than`) sign which translates to `for greater than than x seconds` or `for longer than than x seconds` instead of `for greater than x seconds` or `for longer than x seconds`.