fix bug that num_send is bigger than obs
when I add chinese words to config file and show to the web page,  it always lost something in html-string make webpage shows strange in chrome.
after debug,  I found that num_sent can bigger than obs,  as num_sent is 4206 but obs is 4096. it means  4096 bytes send in the first time, but in the second time, it send from the  position 4207. that's the reason why something lost in the webpage.